+++ opensubdiv/far/CMakeLists.txt
@@ -58,6 +58,7 @@ set(PUBLIC_HEADER_FILES
     patchMap.h
     patchTable.h
     patchTableFactory.h
+    primvarRefiner.h
     ptexIndices.h
     stencilTable.h
     stencilTableFactory.h
