+++ tutorials/far/tutorial_0/far_tutorial_0.cpp
@@ -31,6 +31,7 @@
 //
 
 #include <opensubdiv/far/topologyRefinerFactory.h>
+#include <opensubdiv/far/primvarRefiner.h>
 
 #include <cstdio>
 
@@ -143,8 +144,7 @@ int main(int, char **) {
 
 
     // Interpolate vertex primvar data
-    refiner->Interpolate(verts, verts + nCoarseVerts);
-
+    Far::PrimvarRefiner(*refiner).Interpolate(verts, verts + nCoarseVerts);
 
 
     { // Output OBJ of the highest level refined -----------
